Create Pick Up Ball Functions
The function is called when it reaches the intersection that is just before the ball on the wall. It carefully tracks the line, monitors its distance from the wall, and carefully gets into position to pick up the ball. After the ball function is called, it does a 180 degree turn then returns to the intersection it started at. The function will break once it reaches that intersection. This code includes that following functions:
void pick_up_ball( );
void trackline( );
void forward_straight( );
void veer( );
void stop_robot();
void turn( );
void print_sensor_vals( );