Update Compiled_code_main_program.ino
Debugged the turn function so that it would actually turn in one direction instead of doing a half turn one way, then a half turn back. Cleaned up some of the spacing in functions a bit for readability. Took out the prn_logic parameter from the function that prints line sensor values. Made it so that it detects walls with the bumpers instead of the IR sensor. Have not had any success with testing (although only minimal testing has been done on it) the drop ball portion of the code.